docs: mark unified local server architecture as superseded
